0.2.0 update

+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/actors/mobs/NewTengu.java
@@ -36,7 +36,9 @@ import com.shatteredpixel.shatteredpixeldungeon.actors.buffs.Doom;
 import com.shatteredpixel.shatteredpixeldungeon.actors.buffs.LockedFloor;
 import com.shatteredpixel.shatteredpixeldungeon.actors.buffs.Terror;
 import com.shatteredpixel.shatteredpixeldungeon.actors.hero.Hero;
+import com.shatteredpixel.shatteredpixeldungeon.actors.hero.HeroClass;
 import com.shatteredpixel.shatteredpixeldungeon.actors.hero.HeroSubClass;
+import com.shatteredpixel.shatteredpixeldungeon.actors.hero.Talent;
 import com.shatteredpixel.shatteredpixeldungeon.effects.BlobEmitter;
 import com.shatteredpixel.shatteredpixeldungeon.effects.CellEmitter;
 import com.shatteredpixel.shatteredpixeldungeon.effects.FloatingText;
@@ -198,8 +200,12 @@ public class NewTengu extends Mob {
 	@Override
 	public void die( Object cause ) {
 		
-		if (Dungeon.hero.subClass == HeroSubClass.NONE) {
+		if (Dungeon.hero.subClass == HeroSubClass.NONE && Dungeon.hero.heroClass != HeroClass.ADVENTURER) {
 			Dungeon.level.drop( new TomeOfMastery(), pos ).sprite.drop();
+		}
+		if (Dungeon.hero.heroClass == HeroClass.ADVENTURER){
+			Dungeon.hero.subClass = HeroSubClass.NOTHING_1;
+			Talent.initSubclassTalents(Dungeon.hero);
 		}
 		
 		GameScene.bossSlain();
